requirements.txt
fastapi==0.79.0
uvicorn[standard]==0.18.2
asyncpg==0.27.0
tortoise-orm==0.19.2
aerich==0.7.1
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest==7.1.2